requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
PyMySQL
pydantic
pydantic-settings
passlib[bcrypt]
requests
python-dotenv
httpx
mysql-connector-python